Cleanup of the Debian/Ubuntu packaging scripts
git-svn-id: https://rpg.hamsterrepublic.com/source/wip@1271 7d344553-34f0-0310-a9b1-970ce8f1c3a2

+++ linux/all.sh
@@ -1,8 +1,17 @@
 #!/bin/sh
 rm *.deb
-rm -R ohrrpgce
+
+if [ -d ohrrpgce ] ; then
+  rm -R ohrrpgce
+fi
 ./ohrrpgce.py
-rm -R wandering-hamster
+
+if [ -d wandering-hamster ] ; then
+  rm -R wandering-hamster
+fi
 ./wander.py
-rm -R vikings-of-midgard
+
+if [ -d vikings-of-midgard ] ; then
+  rm -R vikings-of-midgard
+fi
 ./viking.py
